Move chunkifier classes to separate module
